Rename 'clientapi' package to 'api'
